Azure-AI/Creating_FAISS_Vector_search_index/requirements.txt
sentence_transformers
faiss-cpu
numpy
